SQL: Mob_Droplist support for 1000 rolls (allows the use of decimal droprate)

+++ src/map/utils/itemutils.h
@@ -45,7 +45,7 @@ struct DropItem_t
 {
 	uint16 ItemID;
 	uint8  DropType;
-	uint8  DropRate;
+	float  DropRate;
 };
 
 struct LootItem_t
